+++ CHANGELOG.md
@@ -12,6 +12,9 @@ how this file is formatted and how the process works.
 ### Fixed
 - KigoLegacy metadata worker typo
 
+### Added
+- Ability to call `skip_property` from `start`'s block
+
 ## [0.12.2] - 2016-10-07
 ### Added
 - Host id and property id in kigo legacy sync process to Rollbar scope
@@ -27,9 +30,6 @@ how this file is formatted and how the process works.
 
 ### Added
 - SAW: parse owner information for Roomorama::Property
-
-### Added
-- Ability to call `skip_property` from `start`'s block
 
 ## [0.12.0] - 2016-10-05
 ### Added